add error handling

+++ src/main/java/com/manjula/todo/service/TodoServiceImpl.java
@@ -1,6 +1,7 @@
 package com.manjula.todo.service;
 
 import com.manjula.todo.dto.TodoDto;
+import com.manjula.todo.exception.ResourceNotFoundException;
 import com.manjula.todo.model.Todo;
 import com.manjula.todo.repository.TodoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,22 +25,36 @@ public class TodoServiceImpl implements TodoService {
     @Override
     public TodoDto findById(Long id) {
         Todo found = todoRepository.findOne(id);
-        return (found == null) ? null : found.view();
+        if (found == null) {
+            throw new ResourceNotFoundException(id, "Todo not found");
+        }
+        return found.view();
     }
 
     @Override
-    public void update(TodoDto todoDto) {
-        todoRepository.save(Todo.valueOf(todoDto));
+    public void update(Long id, TodoDto todoDto) {
+        TodoDto dto = findById(id);
+        dto.setId(id);
+        dto.setTitle(todoDto.getTitle());
+        dto.setSummary(todoDto.getSummary());
+        todoRepository.save(Todo.valueOf(dto));
     }
 
     @Override
     public void delete(Long id) {
+        Todo found = todoRepository.findOne(id);
+        if (found == null) {
+            throw new ResourceNotFoundException(id, "Todo not found");
+        }
         todoRepository.delete(id);
     }
 
     @Override
     public List<TodoDto> findAll() {
         List<Todo> todos = todoRepository.findAll();
+        if (todos.isEmpty()) {
+            throw new ResourceNotFoundException(null, "Todos not found");
+        }
         return todos.stream().map(Todo::view).collect(Collectors.toList());
     }
